+++ webviz_subsurface/plugins/_rft_plotter/_views/_map_view/_utils/_map_figure.py
@@ -11,7 +11,7 @@ class MapFigure:
         self._ertdf = (
             ertdf.loc[(ertdf["ENSEMBLE"] == ensemble) & (ertdf["ZONE"].isin(zones))]
             .groupby(["WELL", "DATE", "ENSEMBLE"])
-            .aggregate("mean")
+            .mean(numeric_only=False)
             .reset_index()
         )